+++ roxctl/central/export/deployments/serialized_test_deployment.json
@@ -1,0 +1,65 @@
+{
+	"name": "nginx_server",
+	"id": "deaaaaaa-bbbb-4011-0000-111111111111",
+	"clusterId": "caaaaaaa-bbbb-4011-0000-111111111111",
+	"clusterName": "prod cluster",
+	"namespace": "stackrox",
+	"annotations": {
+		"team": "stackrox"
+	},
+	"labels": {
+		"com.docker.stack.namespace": "prevent",
+		"com.docker.swarm.service.name": "prevent_sensor",
+		"email": "vv@stackrox.com",
+		"owner": "stackrox"
+	},
+	"podLabels": {
+		"app": "nginx"
+	},
+	"containers": [
+		{
+			"name": "nginx110container",
+			"image": {
+				"id": "sha256:SHA1",
+				"name": {
+					"registry": "docker.io",
+					"remote": "library/nginx",
+					"tag": "1.10"
+				}
+			},
+			"securityContext": {
+				"addCapabilities": ["SYS_ADMIN"],
+				"dropCapabilities": ["SYS_MODULE"],
+				"privileged": true
+			},
+			"resources": {
+				"cpuCoresRequest": 0.9
+			},
+			"config": {
+				"env": [
+					{
+						"key": "envkey",
+						"value": "envvalue"
+					}
+				]
+			},
+			"volumes": [
+				{
+					"destination": "/vol2",
+					"name": "vol1",
+					"readOnly": true,
+					"source": "/vol1",
+					"type": "host"
+				}
+			],
+			"secrets": [
+				{
+					"name": "secretname",
+					"path": "/var/lib/stackrox"
+				}
+			]
+		}
+	],
+	"created": "2020-12-24T23:59:59.999999999Z",
+	"priority": "1"
+}
